fix: begun fixing merge comments
